Start initial version, with multi-platform support.
Five rendering targets are available:

- Web
- Desktop (Electron)
- Desktop (NWJS)
- iOS
- Android

I have not been able to test how well Android is working with this
setup, so it might require additional work and testing.

The current "app" is just a Hello World starting point, which makes the
boilerplate for each target look highly duplicated. These duplications
are likely to go away once the app grows features, but if not we can
decide how best to streamline.