Fix RTCView idiosyncrasies
1. RTCView doesn't do anything with audio so it's much easier at the
time of this writing to not render the Audio component. Otherwise, we
get into source code duplication without reason.

2. The implementations of RTCView differ between Android and iOS with
respect to the value of streamURL: Android doesn't treat an empty string
the same way as null/undefined like iOS. At the time of this writing it
is much easier to deal with that difference in the Video component.