Remove the static directory
We don't have anything in the static directory but favicon.ico. But 'npm
run start:web' doesn't find favicon.ico in the static directory, it
expects it to live in the root of the source tree. And that is where it
live in the binary distribution (created by 'npm run build:web'). And
favicon.ico is there in the jitsi-meet project.