Reduce duplication
@tsareg commented on a commit of mine that a check of the room name was
buggy & he mentioned that he should rather have it like in the
WelcomePage. I thought that then there should be a single function to
use both in my commit & in the WelcomePage. But it turned out that the 2
OS-specific backends of the WelcomePage didn't use a single function &
rather implemented 1 & the same logic twice.