Fix the Conference page without a room name
It was possible to navigate back and forth and land on the Conference
page without a room name. The problem was that the navigate action was
not automatically related to the room name in the Redux state but rather
received the room name as a function argument. Effectively, such a
design left the navigate action not automatically related to the roomSet
action and the dispatcher had to take care of the correctness.