Switch to master of jitsi/react-native-webrtc
The actual work on the fork jitsi/react-native-webrtc of
oney/react-native-webrtc used to live in the branch lib-jitsi-meet.
Master used to mark the point of forking from upstream's master and then
lib-jitsi-meet was based on master. That is unnecessarily complicated,
especially after lib-jitsi-meet got merged into upstream's master and I
got write access to upstream.